add .node-version and update dev instructions in README
Uncertain what node version to use I started at 14 and
worked backwards through 12 and 10 to 8.8.0, the first
node version which let me both install all the
dependencies and run the server locally (npm run dev)
without error.

I noticed that the .package-lock.json file had updated
so to prevent this I updated the README to use the ci
command instead of the install command:

    npm ci

At some point, npm updated the behaviour of the install
command such that it respects .package.json as guidelines
but does not respect .package-lock.json, which means
that the install command will upgrade whatever
dependencies it notices are available.

The ci command, on the other hand, respects the lock
file so asking occasional contributors to use the ci
command means that you will be less likely to receive
.package-lock.json changes in pull requests